fix(runner-agent-bridge): apply round-2 code-review fixes
A second code review of PR #98 surfaced four IMPORTANT findings; this
commit addresses each one. Rust runner test count goes 146 → 147 (one
strengthened regression test); no behavioural change for the
already-tested paths.

1. views/sessions.py — wrap upsert_runner_live_state in try/except.
   The design's "observability ingestion must never break the poll
   path" invariant was not enforced at the call site. A malformed
   snapshot or a transient DB error would have returned 500 to the
   runner and spun its retry loop. Mirrors the existing pattern for
   drain_for_runner_by_id.

2. models.py + migration 0012 — widen approvals_pending to
   PositiveIntegerField. The runner serialises this as u32 with a
   u32::MAX guard sentinel, which would overflow PositiveSmallInteger
   (max 32 767) and raise DataError on save(). Migration 0012 is in
   this same PR (not yet shipped) so the column type is updated in
   place rather than via a follow-up migration.

3. tasks.py reconcile_stalled_runs — per-row try/except around
   finalize_run_terminal. A single bad row was aborting the whole
   sweep; the next 30s tick would re-walk the same poison row and
   never reap the rest. Logs and continues.

4. state.rs — group the seven Mutex<Option<T>> fields into one
   Mutex<ObservabilitySnapshot>. The original shape required
   reset_run_snapshot to enumerate every field by hand; adding a new
   snapshot field would have silently leaked across run boundaries
   with no compile error. Also collapses note_agent_event from three
   separate locks (where a concurrent reader could see a partial
   update) to one. Public API unchanged. Strengthened the rid-change
   test to assert whole-struct equality with Default so future fields
   are automatically covered.

+++ apps/api/pi_dash/runner/migrations/0012_runner_live_state.py
@@ -49,7 +49,7 @@ class Migration(migrations.Migration):
                 ("agent_subprocess_alive", models.BooleanField(blank=True, null=True)),
                 (
                     "approvals_pending",
-                    models.PositiveSmallIntegerField(blank=True, null=True),
+                    models.PositiveIntegerField(blank=True, null=True),
                 ),
                 ("input_tokens", models.BigIntegerField(blank=True, null=True)),
                 ("output_tokens", models.BigIntegerField(blank=True, null=True)),
